Replaced if statements
Replaced the if statements with a while loop with a try clause inside of it.  There is a dictionary with all of the options inside and makes it more readable. Also added a Kelvin to Fahreneheit conversion. Some of the variables were renamed with the first variable being lower case while the second variable is upper case. I made these changes to not confuse myself and increase readability by reducing the if statements.