Replaced if statements
I replaced the if statements with a dictionary and while cause just like in the main.py file.  I also made the Kelvin to Fahrenheit equation the correct equation and I replaced  the comment in line 17.  I assigned the classes to variables Temp1 to Temp3 as the other variables are placeholders for now. I plan on adding the repeat function to this. I made these changes to make code more readable and less complicated.