Added rankine conversions
The rankine conversions are from Rankine to Fahrenheit, Celsius, and
Kelvin. I also added more onto the dictionary for the conversions and I
made the string for Question more readable. I did this by breaking it up
and separating the conversions out. Comments were added to lines 15, 35, 56, and 75 to tell which conversions they are. Makes it more readable and when I come back, I know what they are. A comment was also added to line 8 to explain it more.