Fixes to an issue
The issue was with the f string not being printed and that when the
repeat function was added after that, the console would not go to that function.
The fix was to just print the dictionary on line 98. For the repeat
function, I set the repeat variable to become a new instance of the Temperature
class and then print it. These changes were made I was testing on
repl.it and I needed to make it work better.