Filled in rankine conversions
Added more onto the functions/attributes of the Rankine class. The
Rankine to Fahrenheit, Celsius, and Kelvin conversions are now added and
should work as the conversions were added to the dictionary as well. The
Fahrenheit, Celsius, and Kelvin to Rankine conversions were added as
well to their respective classes. Added some comments on lines 93 and
101-102 (mult-line comment). I added in these changes so that there
would be rankine conversions and the comments were added to sort of
increase clarity of the project.