Added the option for Rankine
I added the Rankine conversions in the Setup function. Now you can pick
the rankine conversions and the other temperature types to Rankine. The
response variable (111-114) now has the rankine conversions and the options
dictionary (lines 124 -128)also has the conversions added as well. The
output on the console should now show 4 rows of 3 conversions as I've
changed the new lines (\n) as well.I forgot to add it as I thought I did and that's why I made these changes.