Replaced if statements in Repeat()
Added a while loop with a try, except clause to replace the if statements in the Repeat function. I thought I did it on this file but I guess it was in the OOP.py file.  The dictionary should now print the result on line 113 as I forgot to add print() around the options dictionary.