Fixed Error on line 14
I forgot to close the parentheses around print() on line 14 for main.py. I also did some spacing for some the lines as it would be more readable and I also removed some of the parentheses for line 27. These changes are mainly cosmetic and shouldn't change any of the code. I made these changes for readability and for bug fixing.