Changed from degrees Kelvin to Kelvin
For OOP.py, I changed degrees Kelvin to Kelvin as that is more correct.
I also added the type of temperature and I guess I forgot to do that. In
order to change from degrees Kelvin to Kelvin, I had to create
Kelvin_input in the Kelvin class and it asks How much Kelvin instead of
How many degrees. I am also in the process of adding a types to the
Temperature Class so that I don't have to type out each one every single
time. I made these changes as I want to make the Temperature converter
as correct as I can do it.