Add on to previous commit
I forgot to save and this should bring most of the changes I was talking
about such as changing the input in Kelvin and adding the kind of
temperature after degrees. I also forgot to say that I changed the
comment on lines 131- 132 as I think it is more accurate now.