Removed types method
I decided to remove it as I didn't want to deal with accessing variables
inside of a class method. I also moved the comment on line 134 or 135
onto line 17 as that comment should be put there instead. I made these
changes as I want the program to not have lines that don't do anything
or don't work.