Changed celsius to centigrade
This change is for main.py .I changed temperatures that say celsius to centigrade to make it consistent. I only want it to say centigrade instead of having celsius and centigrade mixed together. I made this change so that the temperature type is named the same and even though celsius is the same thing as centigrade, I want it to be named centigrade.