Change temperature print statements
I changed the print statements of the temperature functions to f
strings. This makes it better and I can change the variables better. So
that it says for example 32 degrees fahrenheit (the temperature about to
be converted) to 0 degrees celsius (the temperature you want converted).
The variable temp_type is global and is used for the temperature type
that is about to be converted. I also changed '' to "" just for
organization and to be consistent. There is no difference between '' and
"" in Python. I made these changes after checking it and realizing that
I can make these changes.